Fix AS errors
