+++ smart_contracts/contracts_as/test/set-action-thresholds/assembly/tsconfig.json
